Fix RangeError bug when a .skp file is reopened after being associated with an .idf file. (Issue #13)
The bug resulted from a common mistake where new users would try to save the EnergyPlus model by using SketchUp's normal File/Save command, and then subsequently try to reopen the .skp file. This fix changes the process to never associate the .idf file with the .skp file in the first place. It also adds a warning that is shown when the user tries to save the .skp file to remind them to use Euclid's own Save and Open commands.