Fix check-for-update mechanism to use built-in SketchUp::Http::Request for SU 2017 and later; revert to old UI::WebDialog trick for versions before 2017.
The standard Ruby library Net:HTTP that was used to check for updates (beginning in Euclid 0.9.1) is actually synchronous/blocking and was causing an annoying delay that froze the UI for a few seconds during start up. The new built-in SketchUp::Http::Request call (or UI::WebDialog trick) is asynchronous and eliminates the slow start. The format for the latest-version payload was also reverted to work with UI::WebDialog again.